src/requirements.txt
Django==1.8.*

# Django CMS requirements
django-cms~=3.2.0

# Optional Django CMS dependencies
djangocms-text-ckeditor~=2.8.1
djangocms-link
djangocms-snippet
django-filer~=1.0.6
cmsplugin-filer~=1.0.0
Pillow
django-reversion~=1.9.0
djangocms-bootstrap3
cmsplugin-bootstrap-carousel
djangocms-cascade
git+git://github.com/MagicSolutions/django-orderedmodel.git@0.1.5#egg=django-orderedmodel-0.1.5
git+git://github.com/raphaelm/cmsplugin-carousel.git#egg=cmsplugin-carousel